[serve] move replica info classes to prepare for download_replica_logs method (avoid circular import)
